+++ modules/mpi/common/MPICommon.h
@@ -10,7 +10,7 @@
 #define OMPI_SKIP_MPICXX 1
 #include <mpi.h>
 
-#include "ospcommon/common.h"
+#include "rkcommon/common.h"
 #include "ospray/OSPEnums.h"
 
 #ifdef _WIN32
@@ -47,7 +47,7 @@
 #define OSPRAY_WORLD_GROUP_TAG 290374
 
 namespace mpicommon {
-using namespace ospcommon;
+using namespace rkcommon;
 
 /*! global variable that turns on logging of MPI communication
   (for debugging) _may_ eventually turn this into a real logLevel,
@@ -130,7 +130,7 @@ struct OSPRAY_MPI_INTERFACE Message
   /*! @} */
 
   /*! @{ actual payload of this message */
-  ospcommon::byte_t *data{nullptr};
+  rkcommon::byte_t *data{nullptr};
   size_t size{0};
   /*! @} */
   // TODO WILL: Profiling info, when this message started sending
@@ -144,7 +144,7 @@ struct UserMemMessage : public Message
 {
   UserMemMessage(void *nonCopyMem, size_t size) : Message()
   {
-    data = (ospcommon::byte_t *)nonCopyMem;
+    data = (rkcommon::byte_t *)nonCopyMem;
     this->size = size;
   }