+++ modules/mpi/ospray/render/distributed/AlphaCompositeTileOperation.cpp
@@ -67,7 +67,7 @@ void LiveAlphaCompositeTile::process(const ospray::Tile &tile)
 {
   std::lock_guard<std::mutex> lock(mutex);
   {
-    auto addTile = ospcommon::make_unique<BufferedTile>();
+    auto addTile = rkcommon::make_unique<BufferedTile>();
     std::memcpy(&addTile->tile, &tile, sizeof(tile));
 
     bufferedTiles.push_back(std::move(addTile));